+++ lib/features/digital_guide/presentation/widgets/digital_guide_nav_link.dart
@@ -1,7 +1,6 @@
 import "package:flutter/material.dart";
 
 import "../../../../config/ui_config.dart";
-
 import "../../../../theme/app_theme.dart";
 
 class DigitalGuideNavLink extends StatelessWidget {
