+++ lib/features/home_view/widgets/departments_section/deparment_home_tile.dart
@@ -1,7 +1,6 @@
 import "package:flutter/material.dart";
 
 import "../../../../config/ui_config.dart";
-
 import "../../../../theme/app_theme.dart";
 import "../../../../widgets/my_cached_image.dart";
 import "../../../../widgets/optimized_directus_image.dart";
