+++ lib/features/parkings/parkings_view/widgets/parking_wide_tile_card.dart
@@ -1,7 +1,6 @@
 import "package:flutter/material.dart";
 
 import "../../../../config/ui_config.dart";
-
 import "../../../../theme/app_theme.dart";
 import "../../../../theme/iparking_theme.dart";
 import "../../../../utils/context_extensions.dart";
