+++ lib/features/science_club/science_clubs_filters/widgets/filter_chip.dart
@@ -2,7 +2,6 @@ import "package:flutter/material.dart";
 import "package:flutter_riverpod/flutter_riverpod.dart";
 
 import "../../../../config/ui_config.dart";
-
 import "../../../../theme/app_theme.dart";
 
 class MyFilterChip extends ConsumerWidget {
